requirements.txt
Flask==2.3.3
tensorflow==2.18.0
Pillow>=10.2.0
numpy>=1.26.0
werkzeug==2.3.7
gunicorn==21.2.0
scikit-learn